Simplifies query string. Moves component to folder. Removes mock from test.
